Refactor Constants class package
