+++ app/src/main/java/com/andela/helpmebuy/activities/CreateTravelActivity.java
@@ -21,7 +21,7 @@ import com.andela.helpmebuy.models.Address;
 import com.andela.helpmebuy.models.Location;
 import com.andela.helpmebuy.models.Travel;
 import com.andela.helpmebuy.models.User;
-import com.andela.helpmebuy.utilities.Constants;
+import com.andela.helpmebuy.config.Constants;
 import com.andela.helpmebuy.utilities.CurrentUserManager;
 
 import org.joda.time.DateTime;
