+++ app/src/main/java/com/andela/helpmebuy/activities/HomeActivity.java
@@ -34,7 +34,7 @@ import com.andela.helpmebuy.dal.firebase.FirebaseCollection;
 import com.andela.helpmebuy.models.Location;
 import com.andela.helpmebuy.models.Travel;
 import com.andela.helpmebuy.models.User;
-import com.andela.helpmebuy.utilities.Constants;
+import com.andela.helpmebuy.config.Constants;
 import com.andela.helpmebuy.utilities.CurrentUserManager;
 import com.andela.helpmebuy.utilities.HomeCountryDetector;
 import com.andela.helpmebuy.utilities.HomeCountryDetectorListener;
