+++ app/src/main/java/com/andela/helpmebuy/locations/FirebaseCities.java
@@ -2,7 +2,7 @@ package com.andela.helpmebuy.locations;
 
 import com.andela.helpmebuy.dal.firebase.FirebaseCollection;
 import com.andela.helpmebuy.models.City;
-import com.andela.helpmebuy.utilities.Constants;
+import com.andela.helpmebuy.config.Constants;
 
 public class FirebaseCities extends FirebaseCollection<City> {
     public FirebaseCities(String regionID) {
