+++ app/src/main/java/com/andela/helpmebuy/locations/FirebaseCountries.java
@@ -2,7 +2,7 @@ package com.andela.helpmebuy.locations;
 
 import com.andela.helpmebuy.dal.firebase.FirebaseCollection;
 import com.andela.helpmebuy.models.Country;
-import com.andela.helpmebuy.utilities.Constants;
+import com.andela.helpmebuy.config.Constants;
 
 public class FirebaseCountries extends FirebaseCollection<Country> {
     public FirebaseCountries() {
