+++ app/src/main/java/com/andela/helpmebuy/locations/FirebaseRegions.java
@@ -2,7 +2,7 @@ package com.andela.helpmebuy.locations;
 
 import com.andela.helpmebuy.dal.firebase.FirebaseCollection;
 import com.andela.helpmebuy.models.Region;
-import com.andela.helpmebuy.utilities.Constants;
+import com.andela.helpmebuy.config.Constants;
 
 public class FirebaseRegions extends FirebaseCollection<Region> {
